Ticket history brief and list scroll together
First, scrollview's bottom, otherwise it has no height.
Second, according to
https://doc.qt.io/qt-5.12/qml-qtquick-controls2-scrollview.html#sizing
If contains more than 1, should set contentHeight and contentWidth (even though this works with only setting contentHeight)
Third, I set clip to true because of the same doc page.
"If it is not used as a full-screen item, you should consider setting the clip property to true"
Adjust accordingly.
Fourth, ListView's interactive default *is* true, and what we want here is actually false, and let the ScrollView does the scrolling.